Use pseudo-absolute imports for compatibility with importmap-rails
ScrimController could not be imported in production environments.

Debugging revealed that:
* ./scrim_controller becomes /assets/controllers/scrim_controller
* /assets/controllers/scrim_controller is not present in importmap
* /assets/controllers/scrim_controlller.js cannot be loaded because it
  does not include an sprockets hash.

Changing import to use "controllers/scrim_controller" makes sure that
it matches the importmap syntax but requires changing node and rollup
configuration so that rollup can find our files as modules relative to
a custom root.

Useful reference: https://stackoverflow.com/questions/70548841/how-to-add-custom-js-file-to-new-rails-7-project/72855705#72855705