Turbo 8 compatibility
This is a breaking change – Turbo 8 requires a full HTML response to
Frame requests, and the response head element needs to have the same
data-turbo-track items as the current page or the frame will not be
able to advance history.

This change introduces a replacement for the default layout lambda
that Turbo uses to achieve this, which will return a kpop/frame
layout automatically for kpop frame requests.

This is a breaking change because the layout needs to include all
of the tracked header elements from the application, so we are
changing the name of the layout to help ensure this happens.