Rewrite with HTML5 dialogs
This is pretty-much a re-write of KPOP from the ground up to use HTML5 dialogs.
We no longer use turbo-frame advance to push the modal URL into the browser,
and we don't modify history. Instead, we use the native dialog behaviour to
manage focus and rendering. Scrim is gone, in favour of the dialog's backdrop,
and the server-side integration now lets us write 'normal' actions that work
with both kpop frames and native sheets.